# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioFilter": {
        "enum": [
          "any",
          "with",
          "without"
        ],
        "title": "AudioFilter",
        "type": "string"
      },
      "AutoExclusionUpdate": {
        "properties": {
          "bookmarkIds": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Bookmarkids",
            "type": "array"
          },
          "excluded": {
            "title": "Excluded",
            "type": "boolean"
          }
        },
        "required": [
          "bookmarkIds",
          "excluded"
        ],
        "title": "AutoExclusionUpdate",
        "type": "object"
      },
...
      "BookmarkIds": {
        "properties": {
          "bookmarkIds": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Bookmarkids",
            "type": "array"
          }
        },
        "required": [
          "bookmarkIds"
        ],
        "title": "BookmarkIds",
        "type": "object"
      },
...
      "BookmarkPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Bookmark"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ],
        "title": "BookmarkPage",
        "type": "object"
      },
      "CountResult": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "CountResult",
        "type": "object"
      },
      "ExclusionFilter": {
        "enum": [
          "any",
          "excluded",
          "included"
        ],
        "title": "ExclusionFilter",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobIds": {
        "properties": {
          "jobIds": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Jobids",
            "type": "array"
          }
        },
        "required": [
          "jobIds"
        ],
        "title": "JobIds",
        "type": "object"
      },
...
      "ReadeckActionResult": {
        "description": "How a bulk Readeck action went; a failed bookmark is left as it was.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          }
        },
        "required": [
          "count",
          "failed"
        ],
        "title": "ReadeckActionResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Readeck Audiobook",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/audio/{filename}": {
      "get": {
        "operationId": "download_audio_api_audio__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Audio"
      }
    },
...
    "/api/bookmarks": {
      "get": {
        "operationId": "list_bookmarks_api_bookmarks_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audio",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AudioFilter",
              "default": "any"
            }
          },
          {
            "in": "query",
            "name": "autoGeneration",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ExclusionFilter",
              "default": "any"
            }
          },
          {
            "in": "query",
            "name": "addedFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Addedfrom"
            }
          },
          {
            "in": "query",
            "name": "addedTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Addedto"
            }
          },
          {
            "in": "query",
            "name": "publishedFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedfrom"
            }
          },
          {
            "in": "query",
            "name": "publishedTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publishedto"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bookmarks"
      }
    },
    "/api/bookmarks/audio/delete": {
      "post": {
        "operationId": "delete_bookmark_audio_api_bookmarks_audio_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bookmark Audio"
      }
    },
    "/api/bookmarks/auto-generation": {
      "put": {
        "operationId": "set_auto_exclusion_api_bookmarks_auto_generation_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoExclusionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Auto Exclusion"
      }
    },
    "/api/bookmarks/readeck/archive": {
      "post": {
        "description": "Mark read and archive in Readeck, then drop the bookmark and its audio here.",
        "operationId": "archive_in_readeck_api_bookmarks_readeck_archive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadeckActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive In Readeck"
      }
    },
    "/api/bookmarks/readeck/delete": {
      "post": {
        "description": "Delete from Readeck, then drop the bookmark and its audio here.",
        "operationId": "delete_in_readeck_api_bookmarks_readeck_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadeckActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete In Readeck"
      }
    },
    "/api/jobs": {
      "get": {
        "description": "Queued, running and failed jobs. A completed job is its bookmark's audio.",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs"
      },
      "post": {
        "operationId": "create_jobs_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Jobs"
      }
    },
    "/api/jobs/bulk-delete": {
      "post": {
        "operationId": "bulk_delete_jobs_api_jobs_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobIds"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Delete Jobs"
      }
    },
...
    "/api/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job"
      },
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job"
      }
    },
    "/api/settings": {
      "get": {
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings"
      },
      "put": {
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings"
      }
    },
...
  }
}